Only install camera package from TC PPA if it's not installed
Sicne Pulse19 release, Canonical started to include
qcom-chicdk-qcm6490-rubikpi3 pkg as camera support for Rubik Pi 3.

CAMERA_SETTINGS is also included in qcom-chicdk-qcm6490-rubikpi3.
So there's no need to create it by this script.

(cherry picked from commit b24ed1c74a1fc3ff13e8018ce3fb02c44cf948a6)

+++ install_ppa_pkgs.sh
@@ -152,8 +152,6 @@ install_cam_ai_samples()
 	sudo chown -R ubuntu /opt
 	grep -qxF "$XDG_EXPORT" $USER_HOME/.bashrc || echo "$XDG_EXPORT" >> $USER_HOME/.bashrc
 	sudo bash -c "grep -qxF '${XDG_EXPORT}' /root/.bashrc || echo '${XDG_EXPORT}' >> /root/.bashrc"
-	sudo mkdir -p /var/cache/camera
-	sudo sh -c "echo 'enableNCSService=FALSE' > $CAMERA_SETTINGS"
 	add_cam_ai_pkgs
 }
 
@@ -226,7 +224,11 @@ install()
 {
 	sudo apt install -y ${PKG_LIST[@]}
 	sudo ln -sf libOpenCL.so.1 /usr/lib/aarch64-linux-gnu/libOpenCL.so # Fix for snpe-tools
-	sudo apt install -y rubikpi3-cameras
+    if ! dpkg -s qcom-chicdk-qcm6490-rubikpi3 >/dev/null 2>&1; then
+		sudo mkdir -p /var/cache/camera
+		sudo sh -c "echo 'enableNCSService=FALSE' > $CAMERA_SETTINGS"
+		sudo apt install -y rubikpi3-cameras
+	fi
 	sudo apt upgrade -y
 }
